Public v0.7.1.5

+++ TFF/Assets/scripts/misc/SecretLevelEntryCheck.cs
@@ -4,7 +4,7 @@ using UnityEngine.SceneManagement;
 
 public class SecretLevelEntryCheck : MonoBehaviour
 {
-    string textVal;
+    public string textVal;
     int number;
     public string[] sceneNames; 
 
@@ -13,21 +13,26 @@ public class SecretLevelEntryCheck : MonoBehaviour
         if (File.Exists(Application.dataPath + "\\CoreExe\\CoreTxt.txt"))
             textVal = File.ReadAllText(Application.dataPath + "\\CoreExe\\CoreTxt.txt");
         else
-        {
-            File.Create(Application.dataPath + "\\CoreExe\\CoreTxt.txt");
             RepeatingAction();
-        }
 
         if (int.TryParse(textVal, out number))
             int.TryParse(textVal, out number);
         else
-            File.WriteAllText(Application.dataPath + "\\CoreExe\\CoreTxt.txt", "0");
+            File.WriteAllText(Application.dataPath + "\\CoreExe\\CoreTxt.txt", "0");      
+    }
 
-        switch (number) 
+    void Update()
+    {
+        switch (number)
         {
+            case 0:
+
+                break;
+
             case 1:
                 RepeatingAction();
-                SceneManager.LoadScene(sceneNames[number - 1]);              
+                SceneManager.LoadScene(sceneNames[0]);
+                FindObjectOfType<FileCreator>().Check();
                 break;
 
             default:
@@ -38,10 +43,18 @@ public class SecretLevelEntryCheck : MonoBehaviour
 
     void RepeatingAction()
     {
-        FileInfo fileInfo = new FileInfo(Application.dataPath + "\\CoreExe\\CoreTxt.txt");
-        fileInfo.Attributes = FileAttributes.Normal;
-        File.WriteAllText(Application.dataPath + "\\CoreExe\\CoreTxt.txt", "0");
-        fileInfo.Attributes = FileAttributes.Hidden;
-        number = 0;
+        if(File.Exists(Application.dataPath + "\\CoreExe\\CoreTxt.txt"))
+        {
+            FileInfo fileInfo = new FileInfo(Application.dataPath + "\\CoreExe\\CoreTxt.txt");
+            fileInfo.Attributes = FileAttributes.Normal;
+            File.WriteAllText(Application.dataPath + "\\CoreExe\\CoreTxt.txt", "0");
+            fileInfo.Attributes = FileAttributes.Hidden;
+            number = 0;
+        }
+        else
+        {
+            File.Create(Application.dataPath + "\\CoreExe\\CoreTxt.txt");
+            RepeatingAction();
+        }    
     }
 }
